+++ src/app/app.module.ts
@@ -17,14 +17,13 @@ import { CacheModule } from './services/cache/cache.module';
 import { MessageModule } from './services/message/message.module';
 
 import { AppComponent } from './app.component';
-import { MapEditorComponent } from './components/map-editor/map-editor.component';
-import { MapEditorModule } from './components/map-editor/map-editor.module';
+import { PathFindingDemoComponent, PathFindingDemoModule } from './components/path-finding-demo/path-finding-demo.module';
 
 const appRoutes: Routes = [
   { path: 'admin', component: HomeComponent },
   { path: 'user', component: HomeComponent },
   { path: 'login', component: LoginComponent },
-  { path: 'map', component: MapEditorComponent },
+  { path: 'path-finding', component: PathFindingDemoComponent },
   { path: '',   redirectTo: 'login', pathMatch: 'full' }
 ];
 
@@ -40,6 +39,7 @@ const appRoutes: Routes = [
       }
     ),
     AccountInfoModule,
+    PathFindingDemoModule,
     BrowserAnimationsModule,
     BrowserModule,
     CacheModule,
@@ -50,8 +50,7 @@ const appRoutes: Routes = [
     LoginModule,
     OrderHistoryModule,
     MessageModule,
-    PlaceOrderModule,
-    MapEditorModule
+    PlaceOrderModule
   ]
 })
 export class AppModule { }