+++ src/app/components/account-info/account-info.module.ts
@@ -5,9 +5,11 @@ import { FormsModule } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
 import { MatSelectModule } from '@angular/material/select';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
 
 import { CacheService } from '../../services/cache/cache.service';
-import { MessageService } from '../../services/message/message.service';
+import { ConfirmationModalComponent } from '../confirmation-modal/confirmation-modal.component';
+import { ConfirmationModalModule } from '../confirmation-modal/confirmation-modal.module';
 
 import { AccountInfoComponent } from './account-info.component';
 
@@ -15,22 +17,26 @@ export { AccountInfoComponent } from './account-info.component';
 
 @NgModule({
   imports: [
+    ConfirmationModalModule,
     CommonModule,
     FlexLayoutModule,
     FormsModule,
     MatButtonModule,
     MatSelectModule,
-    MatSlideToggleModule
+    MatSlideToggleModule,
+    MatSnackBarModule
   ],
   declarations: [
     AccountInfoComponent
+  ],
+  entryComponents: [
+    ConfirmationModalComponent
   ],
   exports: [
     AccountInfoComponent
   ],
   providers: [
-    CacheService,
-    MessageService
+    CacheService
   ]
 })
 export class AccountInfoModule { }