+++ src/app/components/confirmation-modal/confirmation-modal.module.ts
@@ -1,0 +1,25 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { MatButtonModule } from '@angular/material/button';
+import { MatDialogModule } from '@angular/material/dialog';
+
+import { ConfirmationModalComponent } from './confirmation-modal.component';
+
+export { ConfirmationModalComponent } from './confirmation-modal.component';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    FlexLayoutModule,
+    MatButtonModule,
+    MatDialogModule
+  ],
+  declarations: [
+    ConfirmationModalComponent
+  ],
+  exports: [
+    ConfirmationModalComponent
+  ]
+})
+export class ConfirmationModalModule { }