+++ src/app/components/home/home.module.ts
@@ -3,8 +3,9 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatMenuModule } from '@angular/material/menu';
-import { BrowserModule } from '@angular/platform-browser';
 import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { BrowserModule } from '@angular/platform-browser';
 
 import { MediaIfModule } from '../../directives/media-if.module';
 import { AccountInfoModule } from '../account-info/account-info.module';
@@ -30,6 +31,7 @@ export { HomeComponent } from './home.component';
     MatIconModule,
     MatMenuModule,
     MatSidenavModule,
+    MatSnackBarModule,
     MediaIfModule,
     OrderHistoryModule,
     PlaceOrderModule,