+++ src/app/components/incoming-batches/incoming-batches.component.ts
@@ -1,6 +1,7 @@
 import * as _ from 'lodash';
 
 import { Component, OnDestroy } from '@angular/core';
+import { MatDialog, MatSnackBar } from '@angular/material';
 import { Subscription } from 'rxjs/Subscription';
 
 import { Batch } from '../../core/batch';
@@ -11,6 +12,8 @@ import { Order } from '../../core/order';
 import { RobotInfo } from '../../core/robot-info';
 import { CacheService } from '../../services/cache/cache.service';
 import { MessageService } from '../../services/message/message.service';
+import { ConfirmationModalComponent } from '../confirmation-modal/confirmation-modal.component';
+import { IncomingBatchesIdentifier } from '../view-identifiers';
 
 @Component({
   selector: 'ms-incoming-batches',
@@ -24,8 +27,9 @@ export class IncomingBatchesComponent implements OnDestroy { // todo: getting an
   incomingBatchSubscription: Subscription;
   configuredRobots: RobotInfo[];
 
-  constructor(private messageService: MessageService, private cache: CacheService) {
-    this.incomingBatchSubscription = this.cache.getIncomingBatches().subscribe(
+  constructor(private messageService: MessageService, private cache: CacheService,
+              private dialog: MatDialog, private snackBar: MatSnackBar) {
+    this.incomingBatchSubscription = this.cache.getIncomingBatches().subscribe( // todo: error handling?
       (batches: Batch[]) => {
         // this is needed so that bartender can keep track of finished drinks after an update
         this.batches.forEach(batch => {
@@ -36,11 +40,12 @@ export class IncomingBatchesComponent implements OnDestroy { // todo: getting an
         });
         this.batches = batches;
       },
-      error => console.error(error) // todo: do something?
+      (error) => {
+        console.log(error);
+      }
     );
-    this.cache.getEnvironmentDetails().then((envDetails: EnvironmentDetails) => {
-      this.configuredRobots = envDetails.configuredRobots;
-    });
+    this.cache.getEnvironmentDetails()
+      .then((envDetails: EnvironmentDetails) => this.configuredRobots = envDetails.configuredRobots);
   }
 
   updateOrderStatus(order: Order, batch: Batch) {
@@ -51,7 +56,22 @@ export class IncomingBatchesComponent implements OnDestroy { // todo: getting an
   }
 
   sendBatch(batch: Batch) {
-    this.messageService.sendBatch(batch);
+    const modalData = {
+      origin: IncomingBatchesIdentifier,
+      batch: batch
+    };
+    const dialogRef = this.dialog.open(ConfirmationModalComponent, {data: modalData});
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.messageService.sendBatch(batch)
+          .catch(err => {
+            this.snackBar.open('Encountered an error while trying to send the batch.', 'Dismiss', {
+              duration: 30000,
+              panelClass: 'mat-snack-bar-error'
+            });
+          });
+      }
+    });
   }
 
   ngOnDestroy(): void {