+++ src/app/components/place-order/place-order.module.ts
@@ -9,10 +9,13 @@ import { MatChipsModule } from '@angular/material/chips';
 import { MatIconModule } from '@angular/material/icon';
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
 
 import { MediaIfModule } from '../../directives/media-if.module';
 import { CacheService } from '../../services/cache/cache.service';
 import { MessageService } from '../../services/message/message.service';
+import { ConfirmationModalComponent } from '../confirmation-modal/confirmation-modal.component';
+import { ConfirmationModalModule } from '../confirmation-modal/confirmation-modal.module';
 
 import { PlaceOrderComponent } from './place-order.component';
 
@@ -20,6 +23,7 @@ export { PlaceOrderComponent } from './place-order.component';
 
 @NgModule({
   imports: [
+    ConfirmationModalModule,
     CommonModule,
     FlexLayoutModule,
     FormsModule,
@@ -30,11 +34,15 @@ export { PlaceOrderComponent } from './place-order.component';
     MatIconModule,
     MatInputModule,
     MatSelectModule,
+    MatSnackBarModule,
     MediaIfModule,
     ReactiveFormsModule
   ],
   declarations: [
     PlaceOrderComponent
+  ],
+  entryComponents: [
+    ConfirmationModalComponent
   ],
   exports: [
     PlaceOrderComponent