+++ src/app/services/cache/cache.service.ts
@@ -3,7 +3,6 @@ import 'rxjs/add/observable/interval';
 
 import { Injectable, OnDestroy } from '@angular/core';
 
-import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
 import { Subscription } from 'rxjs/Subscription';
 
@@ -25,7 +24,6 @@ export class CacheService implements OnDestroy {
 
   private environmentDetails: EnvironmentDetails;
 
-  // todo: running into issues with the cache when batches at empty, and then order is placed
   private batchUpdatesSubject: Subject<Batch[]>;
   private batchUpdatesCache: Batch[] = [];
 
@@ -152,8 +150,7 @@ export class CacheService implements OnDestroy {
           .then(environmentDetails => {
             this.environmentDetails = environmentDetails;
             resolve(this.environmentDetails);
-          })
-          .catch(error => reject(error));
+          });
       });
     }
   }