+++ src/app/services/message/message.service.ts
@@ -1,200 +1,248 @@
 import { HttpClient } from '@angular/common/http';
-import { Inject, Injectable } from '@angular/core';
+import { Injectable } from '@angular/core';
 import { Subject } from 'rxjs/Subject';
 
 import { environment } from '../../../environments/environment';
 
+import { reject } from 'q';
 import { Batch } from '../../core/batch';
 import { DeliveryLocation } from '../../core/delivery-location';
+import { Edge } from '../../core/edge';
 import { EnvironmentDetails } from '../../core/environment-details';
+import { LocationMap } from '../../core/location-map';
 import { Order } from '../../core/order';
 import { OrderInfo } from '../../core/order-info';
-import { OrderOption } from '../../core/order-option';
 import { SystemDetails } from '../../core/system-details';
 import { UserInfo } from '../../core/user-info';
-import { VerticesAndEdges } from '../../core/vertices-and-edges';
+import { Vertex } from '../../core/vertex';
 
 @Injectable()
 export class MessageService {
+  private static maxNumberOfRetries = 1;
+  private static retryErrorCode = 418;
 
   private sessionKey: string;
-  private user: UserInfo; // todo: would be nice to get rid of this, don't think it's possible
+  private user: UserInfo;
 
   userUpdates: Subject<UserInfo> = new Subject<UserInfo>();
   orderPlacedUpdate: Subject<Order> = new Subject<Order>();
 
   constructor(private http: HttpClient) {}
 
-  getEnvironmentDetails(): Promise<EnvironmentDetails> {
-    return new Promise<any>((resolve, reject) => {
-      const body = {
-        username: this.user.username,
-        sessionKey: this.sessionKey
-      };
-
-      this.http.post(environment.backendUrl + 'getEnvironmentDetails', body)
-        .subscribe((response: EnvironmentDetails) => resolve(response), error => reject(error));
-    });
+  getEnvironmentDetails(retryCount = 0): Promise<EnvironmentDetails> {
+    const body = {
+      username: this.user.username,
+      sessionKey: this.sessionKey
+    };
+    return this.http.post(environment.backendUrl + 'getEnvironmentDetails', body).toPromise()
+      .then((response: EnvironmentDetails) => response)
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.getEnvironmentDetails(retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
-  login(username: string, password: string): Promise<UserInfo> {
-    return new Promise((resolve, reject) => {
-      this.http.post(environment.backendUrl + 'login', {username: username, password: password})
-        .subscribe(
-          (response: {sessionKey: string, userInfo: UserInfo}) => {
-            this.sessionKey = response.sessionKey;
-            this.user = response.userInfo;
-            this.userUpdates.next(this.user);
-            resolve();
-          },
-          error => reject('Invalid username or password') // todo: can do a better job of handling this
-        );
-    });
+  login(username: string, password: string, retryCount = 0): Promise<UserInfo> {
+    return this.http.post(environment.backendUrl + 'login', {username: username, password: password}).toPromise()
+      .then((response: {sessionKey: string, userInfo: UserInfo}) => {
+        this.sessionKey = response.sessionKey;
+        this.user = response.userInfo;
+        this.userUpdates.next(this.user);
+        return this.user;
+      })
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.login(username, password, retryCount + 1);
+        }
+        return Promise.reject(err.status === MessageService.retryErrorCode);
+      });
   }
 
-  logout(): Promise<void> {
-    return new Promise<void>((resolve, reject) => {
-      const body = {username: this.user.username, sessionKey: this.sessionKey};
-      this.http.post(environment.backendUrl + 'logout', body)
-        .subscribe(
-          () => {
-            this.sessionKey = undefined;
-            this.user = undefined;
-            this.userUpdates.next(undefined);
-            resolve();
-          },
-          error => reject(error)
-        );
-    });
+  logout(retryCount = 0): Promise<void> {
+    // todo: make sure all subscriptions are finished
+    // todo: change flag in the service, subs check that flag before next request, once they all give the "OK" resolve to this promise
+    const body = {username: this.user.username, sessionKey: this.sessionKey};
+    return this.http.post(environment.backendUrl + 'logout', body).toPromise()
+      .then(() => {
+        this.sessionKey = undefined;
+        this.user = undefined;
+        this.userUpdates.next(undefined);
+      })
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.logout(retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
-  updateAccountInfo(user: UserInfo): Promise<void> {
-    return new Promise<void>((resolve, reject) => {
-      const body = {
-        username: user.username,
-        sessionKey: this.sessionKey,
-        userInfo: user
-      };
-      this.http.post(environment.backendUrl + 'updateAccountInfo', body)
-        .subscribe(
-          () => {
-            this.user = user;
-            this.userUpdates.next(this.user);
-            resolve();
-          },
-          error => reject(error));
-    });
+  updateAccountInfo(user: UserInfo, retryCount = 0): Promise<void> {
+    const body = {
+      username: user.username,
+      sessionKey: this.sessionKey,
+      userInfo: user
+    };
+    return this.http.post(environment.backendUrl + 'updateAccountInfo', body).toPromise()
+      .then(() => {
+        this.user = user;
+        this.userUpdates.next(this.user);
+      })
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.updateAccountInfo(user, retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
-  sendBatch(batch: Batch): Promise<void> {
-    return new Promise<void>((resolve, reject) => {
-      const body = {
-        username: this.user.username,
-        sessionKey: this.sessionKey,
-        batchId: batch.id
-      };
-
-      // todo: can we just map this???
-      this.http.post(environment.backendUrl + 'sendBatch', body).subscribe(() => resolve(), error => reject(error));
-    });
+  sendBatch(batch: Batch, retryCount = 0): Promise<void> {
+    const body = {
+      username: this.user.username,
+      sessionKey: this.sessionKey,
+      batchId: batch.id
+    };
+    return this.http.post(environment.backendUrl + 'sendBatch', body).toPromise()
+      .then(() => null)
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.sendBatch(batch, retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
-  getIncomingBatches(): Promise<Batch[]> {
-    return new Promise<Batch[]>((resolve, reject) => {
-      const body = {username: this.user.username, sessionKey: this.sessionKey};
-      this.http.post(environment.backendUrl + 'getIncomingBatches', body)
-        .subscribe(
-          (response: {batches: Batch[]}) => resolve(response.batches),
-          error => reject(error)
-        );
-    });
+  getIncomingBatches(retryCount = 0): Promise<Batch[]> {
+    const body = {username: this.user.username, sessionKey: this.sessionKey};
+    return this.http.post(environment.backendUrl + 'getIncomingBatches', body).toPromise()
+      .then((response: {batches: Batch[]}) => {
+        return response.batches;
+      })
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.getIncomingBatches(retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
-  getOrderHistory(): Promise<Order[]> {
-    return new Promise<Order[]>((resolve, reject) => {
-      const body = {username: this.user.username, sessionKey: this.sessionKey};
-      this.http.post(environment.backendUrl + 'getOrderHistory', body).subscribe(
-        (response: {orderHistory: Order[]}) => resolve(response.orderHistory),
-        error => reject(error)
-      );
-    });
+  getOrderHistory(retryCount = 0): Promise<Order[]> {
+    const body = {username: this.user.username, sessionKey: this.sessionKey};
+    return this.http.post(environment.backendUrl + 'getOrderHistory', body).toPromise()
+      .then((response: {orderHistory: Order[]}) => response.orderHistory)
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.getOrderHistory(retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
-  getOrderUpdates(ordersToMonitor: Order[]): Promise<Order[]> {
-    return new Promise<Order[]>((resolve, reject) => {
-      const body = {
-        username: this.user.username,
-        sessionKey: this.sessionKey,
-        orders: ordersToMonitor
-      };
-      this.http.post(environment.backendUrl + 'getOrderUpdates', body)
-        .subscribe(
-          (response: {orders: Order[]}) => resolve(response.orders),
-          error => reject(error)
-        );
-    });
+  getOrderUpdates(ordersToMonitor: Order[], retryCount = 0): Promise<Order[]> {
+    const body = {
+      username: this.user.username,
+      sessionKey: this.sessionKey,
+      orders: ordersToMonitor
+    };
+    return this.http.post(environment.backendUrl + 'getOrderUpdates', body).toPromise()
+      .then((response: {orders: Order[]}) => response.orders)
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.getOrderUpdates(ordersToMonitor, retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
-  // todo: consolidate input arguments down
-  placeOrder(selectedBeverage: OrderOption, selectedAddOns: {key: string, value: string | boolean | number}[],
-             deliveryLocation: DeliveryLocation): Promise<Order> {
-    return new Promise<Order>((resolve, reject) => {
-      const body = {
-        username: this.user.username,
-        sessionKey: this.sessionKey,
-        orderInfo: new OrderInfo(selectedBeverage, selectedAddOns),
-        deliveryLocation: deliveryLocation
-      };
-      this.http.post(environment.backendUrl + 'placeOrder', body)
-        .subscribe((response: {order: Order}) => {
-          this.orderPlacedUpdate.next(response.order);
-          resolve(response.order);
-        }, error => reject(error));
-    });
+  placeOrder(orderInfo: OrderInfo, deliveryLocation: DeliveryLocation, retryCount: number = 0): Promise<Order> {
+    const body = {
+      username: this.user.username,
+      sessionKey: this.sessionKey,
+      orderInfo: orderInfo,
+      deliveryLocation: deliveryLocation
+    };
+    return this.http.post(environment.backendUrl + 'placeOrder', body).toPromise()
+      .then((response: {order: Order}) => {
+        this.orderPlacedUpdate.next(response.order);
+        return response.order;
+      }).catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.placeOrder(orderInfo, deliveryLocation, retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
-  getSystemDetails(): Promise<SystemDetails> {
-    return new Promise<SystemDetails>((resolve, reject) => {
-      const body = {
-        username: this.user.username,
-        sessionKey: this.sessionKey
-      };
-
-      this.http.post(environment.backendUrl + 'getSystemDetails', body).subscribe(
-        (response: SystemDetails) => resolve(response),
-        error => reject(error)
-      );
-    });
+  getSystemDetails(retryCount = 0): Promise<SystemDetails> {
+    const body = {
+      username: this.user.username,
+      sessionKey: this.sessionKey
+    };
+    return this.http.post(environment.backendUrl + 'getSystemDetails', body).toPromise()
+      .then((response: SystemDetails) => response)
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.getSystemDetails(retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
-  getVerticesAndEdges(): Promise<VerticesAndEdges> {
-    return new Promise<VerticesAndEdges>((resolve, reject) => {
-      const body = {
-        username: '',
-        sessionKey: ''
-      };
-
-      this.http.post(environment.backendUrl + 'getVerticesAndEdges', body).subscribe(
-        (response: VerticesAndEdges) => resolve(response),
-        error => reject(error)
-      );
-    });
+  getMap(retryCount = 0): Promise<LocationMap> {
+    return this.http.get(environment.backendUrl + 'getMap').toPromise()
+      .then((response: any) => {
+        const map = new LocationMap();
+        response.vertices.forEach((vertex: any) => {
+          map.vertices.push(new Vertex(vertex));
+        });
+        response.edges.forEach((edge: any) => {
+          map.edges.push(new Edge(edge));
+        });
+        return map;
+      })
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.getMap(retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
-  useAStar(vertices: any): Promise<VerticesAndEdges> {
-    return new Promise<VerticesAndEdges>((resolve, reject) => {
-      const body = {
-        username: '',
-        sessionKey: '',
-        vertexValues: vertices
-      };
+  getPath(vertices: Vertex[], retryCount = 0): Promise<Edge[]> {
+    const body = {
+      username: '',
+      sessionKey: '',
+      vertexValues: vertices
+    };
+    return this.http.post(environment.backendUrl + 'getPath', body).toPromise()
+      .then(((response: any[]) => {
+        const convertedEdges: Edge[] = [];
 
-      // todo: can we just map this???
-      this.http.post(environment.backendUrl + 'useAStar', body).subscribe(
-        (response: VerticesAndEdges) => resolve(response),
-        error => reject(error)
-      );
-    });
+        response.forEach(edge => {
+          convertedEdges.push(new Edge(edge));
+        });
+
+        return convertedEdges;
+      }))
+      .catch(err => {
+        if (err.status === MessageService.retryErrorCode && retryCount < MessageService.maxNumberOfRetries) {
+          retryCount++;
+          return this.getPath(vertices, retryCount + 1);
+        }
+        return Promise.reject(err);
+      });
   }
 
 }