+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# [1.2.0](https://github.com/bcorbold/MobileServerFrontEnd/tree/1.2.0) - Released (25-04-2018)
+- (feat): `PathFindingDemoComponent` updated for pre-calculated paths
+- (feat): Added `ConfirmationModalComponent` for added user validation
+- (feat): `MessageService` now retries requests if failed the first time
+- (style): App now has dark and light theme ready for use
+
 # [1.1.0](https://github.com/bcorbold/MobileServerFrontEnd/tree/1.1.0) - Released (15-04-2018)
 - (feat): `CacheService` now waits until a request has been completed before sending out another request
 - (fix): `CacheService` now emits `Order[]` even if none of them need further updates